requirements.txt
python-dotenv>=1.0.0
elevenlabs>=1.0.0
twilio>=9.0.0
openai-whisper
openai
